libstagefright: use miniz in place of libz
Summary:
Miniz is a lightweight drop in zlib replacement.
Using a basic benchmark while utilizing miniz in libpng
with the equivalent of Z_DEFAULT_COMPRESSION (level 6),
I noted ~8% time decrease in decompression, and ~10%
time decrease in compression. The time decrease was
approximately the same percentage for compression level
2 as well.

Another advantage to uzing miniz for more libraries is the
size decrease from linking to miniz over zlib.

Test Plan: all tests are good to go with this change

Reviewers: randall.rushing, synergy

Reviewed By: synergy

Differential Revision: http://review.codefirex.com/D458